Move `PacketStatus` to `gige::protocol` root
