+++ device/src/gige/protocol/ack.rs
@@ -17,13 +17,16 @@ use crate::gige::{
     Error, Result,
 };
 
+use super::PacketStatus;
+
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub struct AckPacket<'a> {
     header: Header,
     raw_ack_data: &'a [u8],
 }
 impl<'a> AckPacket<'a> {
-    pub fn parse(buf: &'a [u8]) -> Result<Self> {
+    pub fn parse(buf: &'a impl AsRef<[u8]>) -> Result<Self> {
+        let buf = buf.as_ref();
         let mut cursor = io::Cursor::new(buf);
         let header = Header::parse(&mut cursor)?;
 
@@ -60,14 +63,14 @@ impl<'a> AckPacket<'a> {
         self.header.request_id
     }
 
-    pub fn status(&self) -> Status {
+    pub fn status(&self) -> PacketStatus {
         self.header.status
     }
 }
 
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub struct Header {
-    status: Status,
+    status: PacketStatus,
     ack_kind: AckKind,
     length: u16,
     request_id: u16,
@@ -75,7 +78,7 @@ pub struct Header {
 
 impl Header {
     fn parse(cursor: &mut io::Cursor<&[u8]>) -> Result<Self> {
-        let status = Status::parse(cursor)?;
+        let status = PacketStatus::parse(cursor)?;
         let ack_kind = AckKind::parse(cursor)?;
         let length = cursor.read_bytes_be()?;
         let request_id = cursor.read_bytes_be()?;
@@ -90,92 +93,6 @@ impl Header {
 
 pub trait ParseAckData<'a>: Sized {
     fn parse(raw_data: &'a [u8], header: &Header) -> Result<Self>;
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct Status {
-    code: u16,
-    kind: StatusKind,
-}
-
-impl Status {
-    pub fn is_success(self) -> bool {
-        matches!(self.kind, StatusKind::Success | StatusKind::PacketResend)
-    }
-
-    pub fn code(self) -> u16 {
-        self.code
-    }
-
-    pub fn kind(self) -> StatusKind {
-        self.kind
-    }
-
-    fn parse(cursor: &mut io::Cursor<&[u8]>) -> Result<Self> {
-        let code: u16 = cursor.read_bytes_be()?;
-        let kind = match code {
-            0x0000 => StatusKind::Success,
-            0x0100 => StatusKind::PacketResend,
-            0x8001 => StatusKind::NotImplemented,
-            0x8002 => StatusKind::InvalidParameter,
-            0x8003 => StatusKind::InvalidAdderess,
-            0x8004 => StatusKind::WriteProtect,
-            0x8005 => StatusKind::BadAlignment,
-            0x8006 => StatusKind::AccessDenied,
-            0x8007 => StatusKind::Busy,
-            0x8008 => StatusKind::LocalProblem,
-            0x8009 => StatusKind::MessageMismatch,
-            0x800a => StatusKind::InvalidProtocol,
-            0x800b => StatusKind::NoMessage,
-            0x800c => StatusKind::PacketUnavailable,
-            0x800d => StatusKind::DataOverrun,
-            0x800e => StatusKind::InvalidHeader,
-            0x800f => StatusKind::WrongConfig,
-            0x8010 => StatusKind::NotYetAvailable,
-            0x8011 => StatusKind::CurrentAndPrevPacketRemoved,
-            0x8012 => StatusKind::CurrentPacketRemoved,
-            0x8013 => StatusKind::NoReferenceTime,
-            0x8014 => StatusKind::TemporarilyUnavailable,
-            0x8015 => StatusKind::Overflow,
-            0x8016 => StatusKind::ActionLate,
-            0x8fff => StatusKind::GenericError,
-            _ => {
-                return Err(Error::InvalidPacket(
-                    format! {"invalid gige ack status code {:#X}", code}.into(),
-                ));
-            }
-        };
-        Ok(Self { code, kind })
-    }
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum StatusKind {
-    Success,
-    PacketResend,
-    NotImplemented,
-    InvalidParameter,
-    InvalidAdderess,
-    WriteProtect,
-    BadAlignment,
-    AccessDenied,
-    Busy,
-    LocalProblem,
-    MessageMismatch,
-    InvalidProtocol,
-    NoMessage,
-    PacketUnavailable,
-    DataOverrun,
-    InvalidHeader,
-    WrongConfig,
-    NotYetAvailable,
-    CurrentAndPrevPacketRemoved,
-    CurrentPacketRemoved,
-    NoReferenceTime,
-    TemporarilyUnavailable,
-    Overflow,
-    ActionLate,
-    GenericError,
 }
 
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
